feat(config): 支持人类可读字节大小配置（owner_quotas/bucket_limits/vol_capacity） (#327)
新增 internal/size.ParseSize：解析 5GiB/2GB/1.5MiB 等（SI 十进制 + IEC 二进制、大小写不敏感、可带小数/空格），纯数字=字节；owner_quotas/bucket_limits/vol_capacity 改 ByteSize 类型（UnmarshalText 委托 ParseSize），yaml.v3 与 viper 双轨解码（viper 侧 sproxycfg 注册 decode hook 并保留默认 duration/slice hook），旧纯数字配置完全兼容。

+++ pkg/server/config_validate.go
@@ -82,7 +82,7 @@ func (c *Config) Validate() error {
 		}
 		seenRoots[rootKey] = true
 		if v.VolCapacity < 0 {
-			return fmt.Errorf("卷 %q 容量上限 %d 非法：不能为负", v.Name, v.VolCapacity)
+			return fmt.Errorf("卷 %q 容量上限 %d 非法：不能为负", v.Name, int64(v.VolCapacity))
 		}
 		if a := v.ACL; a != nil {
 			if a.Mode != VolumeACLAllow && a.Mode != VolumeACLDeny {
@@ -198,7 +198,7 @@ func (c *Config) Validate() error {
 			return fmt.Errorf("bucket_limits 键 %q 非法：分层配额仅支持 user 桶子目录（如 user/videos/hd），其余功能桶无子目录结构", path)
 		}
 		if limit < 0 {
-			return fmt.Errorf("bucket_limits[%q] 上限 %d 非法：配额上限不能为负", path, limit)
+			return fmt.Errorf("bucket_limits[%q] 上限 %d 非法：配额上限不能为负", path, int64(limit))
 		}
 	}
 	if c.RateLimit.Enabled && c.RateLimit.Requests <= 0 {
